rename : guestBook -> guest-book
