+++ src/views/guest-book/guest-book.ts
@@ -1,5 +1,5 @@
 import Layout from "../../components/Layout/layout";
-import styles from "./guestBook.css";
+import styles from "./guest-book.css";
 export default function GuestBook() {
   const content = `<div class=${styles.guestBookContainer}>방명록 페이지</div>`;
   return Layout(content);
